+++ gha_utils/__init__.py
@@ -17,4 +17,4 @@
 
 from __future__ import annotations
 
-__version__ = "5.6.2"
+__version__ = "5.6.3"