+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [5.6.3 (unreleased)](https://github.com/kdeldycke/workflows/compare/v5.6.2...main)
+
+> [!IMPORTANT]
+> This version is not released yet and is under active development.
+
 ## [5.6.2 (2026-02-02)](https://github.com/kdeldycke/workflows/compare/v5.6.1...v5.6.2)
 
 - Auto-update `gha-utils==X.Y.Z` CLI version pins in workflow files during release.